Change for automatic login from homepage. read more...
If a user has only one account, they will be redirected to that account upon
login from the homepage.
If a user has more than one account, they will be shown a list of accounts.
This might need a different standalone view to be more user-friendly.

Forgery protection reactivated in user_sessions. We are now using global cookies.